repository item /lhuaman

+++ src/main/java/com/proyecto/mc_cotizacion/repository/QuotationRepository.java
@@ -3,5 +3,5 @@ package com.proyecto.mc_cotizacion.repository;
 import com.proyecto.mc_cotizacion.entity.Quotation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuotationRepository extends JpaRepository<Quotation,Integer> {
+public interface QuotationRepository extends JpaRepository<Quotation,Long> {
 }
